# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthCredentials {
  email: string;
  /** @minLength 8 */
  password: string;
}

export type DriverSubscriptionTier = typeof DriverSubscriptionTier[keyof typeof DriverSubscriptionTier];


export const DriverSubscriptionTier = {
  free: 'free',
  pro: 'pro',
} as const;

export interface Driver {
  id: number;
  email: string;
  subscriptionTier: DriverSubscriptionTier;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  driver: Driver;
}

export interface DriverPreferencesInput {
  /** @minimum 0 */
  minFareNgn: number;
  /** @minimum 0.5 */
  maxPickupRadiusKm: number;
  /**
     * @minimum 0
     * @maximum 5
     */
  minPlatformRating?: number | null;
  blacklistedZones: string[];
}

export type DriverPreferences = DriverPreferencesInput & {
  driverId: number;
  updatedAt: string;
};

export type RideRequestStatus = typeof RideRequestStatus[keyof typeof RideRequestStatus];


export const RideRequestStatus = {
  pending: 'pending',
  accepted: 'accepted',
  declined: 'declined',
  expired: 'expired',
} as const;

export interface RideRequest {
  id: number;
  platformSource: string;
  fareAmountNgn: number;
  pickupDistanceKm: number;
  pickupZone: string;
  destinationZone: string;
  status: RideRequestStatus;
  timestamp: string;
}

export type RideDecisionAction = typeof RideDecisionAction[keyof typeof RideDecisionAction];


export const RideDecisionAction = {
  accept: 'accept',
  decline: 'decline',
} as const;

export interface RideDecision {
  action: RideDecisionAction;
}

export interface PlatformConnection {
  id: number;
  platformName: string;
  isActive: boolean;
}