# lib/api-zod/src/generated/types/driver.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DriverSubscriptionTier } from './driverSubscriptionTier';

export interface Driver {
  id: number;
  email: string;
  subscriptionTier: DriverSubscriptionTier;
  createdAt: Date;
}