# lib/api-zod/src/generated/types/driverPreferences.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DriverPreferencesInput } from './driverPreferencesInput';

export type DriverPreferences = DriverPreferencesInput & {
  driverId: number;
  updatedAt: Date;
};