# lib/api-zod/src/generated/types/driverPreferencesInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DriverPreferencesInput {
  /** @minimum 0 */
  minFareNgn: number;
  /** @minimum 0.5 */
  maxPickupRadiusKm: number;
  /**
     * @minimum 0
     * @maximum 5
     */
  minPlatformRating?: number | null;
  blacklistedZones: string[];
}