# lib/api-zod/src/generated/types/rideDecision.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RideDecisionAction } from './rideDecisionAction';

export interface RideDecision {
  action: RideDecisionAction;
}