# lib/api-zod/src/generated/types/rideRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RideRequestStatus } from './rideRequestStatus';

export interface RideRequest {
  id: number;
  platformSource: string;
  fareAmountNgn: number;
  pickupDistanceKm: number;
  pickupZone: string;
  destinationZone: string;
  status: RideRequestStatus;
  timestamp: Date;
}